requirements.txt
crawl4ai
groq==0.11.0
ipykernel==6.29.5
ipython==8.29.0
langchain==0.3.7
langchain-core==0.3.20
langchain-groq==0.2.1
langchain-community==0.3.1
langchain-text-splitters==0.3.2
langgraph==0.2.53
langsmith==0.1.140
httpx==0.27.2
numpy==1.26.4
pandas==2.2.3
pillow==10.4.0
protobuf==5.28.3
pyarrow==18.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
python-dotenv==1.0.1
requests==2.32.3
streamlit==1.39.0
tavily-python==0.5.0
tiktoken==0.8.0
tzdata==2024.2
urllib3==2.2.3
firecrawl-py
linkup-sdk
langchain_linkup
pytest==7.1.3
black
pytest-cov==4.0.0
pylint
mypy
pyright
langchain_tests
pandas-stubs 
ruff
snscrape
selenium==4.22.0
webdriver-manager==4.0.0
undetected-chromedriver==3.5.5
aiohttp==3.11.11
beautifulsoup4
yfinance
unidecode
finta
matplotlib

